src/main/java/dev/thiagorodrigues/livraria/application/controllers/UsuarioController.java
package dev.thiagorodrigues.livraria.application.controllers;

import dev.thiagorodrigues.livraria.application.dtos.UsuarioFormDto;
import dev.thiagorodrigues.livraria.application.dtos.UsuarioResponseDto;
import dev.thiagorodrigues.livraria.domain.services.UsuarioService;
import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import javax.validation.Valid;

@Api(tags = "Usuarios")
@RestController
@RequestMapping("/usuarios")
@RequiredArgsConstructor
public class UsuarioController {

    private final UsuarioService usuarioService;

    @ApiOperation("Criar novo usuário")
    @PostMapping
    public ResponseEntity<UsuarioResponseDto> register(@RequestBody @Valid UsuarioFormDto usuarioFormDto) {
        var usuarioResponseDto = this.usuarioService.criar(usuarioFormDto);

        return ResponseEntity.status(HttpStatus.CREATED).body(usuarioResponseDto);
    }

}
